refactor(error): move UiError to app::error module
Create new app::error module for UI-facing error types and remove
deprecated UiError from crate root.

- Add crates/ploke-tui/src/app/error.rs with UiError enum
- Add pub mod error to app/mod.rs
- Remove deprecated UiError from lib.rs
- Update events.rs to use crate::app::error::UiError
